+++ qserver/instrument/__init__.py
@@ -9,4 +9,4 @@ logger = logging.getLogger(__name__)
 logger.info(__file__)
 print(__file__)
 
-from .iconfig_dict import iconfig
+from ._iconfig_dict import iconfig
