+++ qserver/instrument/devices/simulated_shutter.py
@@ -12,7 +12,6 @@ logger = logging.getLogger(__name__)
 logger.info(__file__)
 print(__file__)
 
-# from ..iconfig_dict import iconfig
 import apstools.devices
 
 shutter = apstools.devices.SimulatedApsPssShutterWithStatus(
@@ -21,5 +20,3 @@ shutter = apstools.devices.SimulatedApsPssShutterWithStatus(
 
 # shutter simulator needs short recovery time after moving
 shutter.delay_s = 0.05
-
-# TODO: suspender?
